refactor login/signup pages
