--- frontend/app/login/login.css
@@ -1,86 +1,0 @@
-.login {
-  background-color: #22324a;
-  display: flex;
-  justify-content: center;
-  padding: 2rem;
-}
-
-.username {
-  padding-top: 1rem;
-  padding-bottom: 1rem;
-  max-width: 300px;
-}
-
-.username input,
-.password input {
-  width: 100%;
-  background-color: #d8c1c1;
-  border-radius: 6px;
-  font-size: 14px;
-}
-
-.password {
-  padding-bottom: 0.4%;
-}
-
-.username-text .password-text {
-  padding-bottom: 0.3%;
-  color: white;
-}
-
-.forgot-password {
-  font-size: 12px;
-  cursor: pointer;
-  text-align: right;
-}
-
-.login-bottom-buttons {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding-bottom: 0.7%;
-  padding-top: 2%;
-}
-
-.login-button {
-  background-color: #191c2d;
-  color: white;
-  border-color: white;
-}
-
-.remember-me-button,
-.create-account-button,
-.login-button {
-  flex: 1;
-  text-align: center;
-}
-
-.create-account-button,
-.login-button {
-  border-radius: 20px;
-  padding: 0.5rem;
-  cursor: pointer;
-  transition:
-    background-color 0.3s,
-    color 0.3s;
-}
-
-.create-account-button:hover,
-.login-button:hover,
-.remember-me-button label:focus,
-.create-account-button:focus,
-.login-button:focus {
-  background-color: #364159;
-}
-
-@media only screen and (max-width: 600px) {
-  .login {
-    padding: 1rem;
-  }
-
-  .username,
-  .password,
-  .login-bottom-buttons {
-    padding: 0.5rem;
-  }
-}
